syncing to head of TFS. Tagging 0.9.3

+++ Merlin/Main/Hosts/SilverLight/Samples1/clean.bat
@@ -1,7 +1,6 @@
 @echo off
 
 echo removing dlr directory, dlr.js, and dlr.xap
-rmdir /S /Q %~dp0dlr > %~dp0temp.log
-del /Q %~dp0dlr.js > %~dp0temp.log
-del /Q %~dp0dlr.xap > %~dp0temp.log
-del %~dp0temp.log
+if exist %~dp0dlr rmdir /S /Q %~dp0dlr > %~dp0temp.log
+if exist %~dp0dlr.js del /Q %~dp0dlr.js > %~dp0temp.log
+if exist %~dp0temp.log del %~dp0temp.log
